chest-xray-benchmark/requirements.txt
# =============================================================================
# Chest X-ray Multi-Disease Detection Benchmark — Requirements
# =============================================================================
# Python 3.10+ required

# ── Core ML / DL ─────────────────────────────────────────────────────────────
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
timm==0.9.16
ultralytics==8.1.34

# ── Classical ML ──────────────────────────────────────────────────────────────
scikit-learn==1.4.1.post1
scikit-image==0.22.0
xgboost==2.0.3
optuna==3.5.0

# ── Data / Image Processing ──────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.1
opencv-python-headless==4.9.0.80
Pillow==10.2.0
albumentations==1.3.1

# ── Visualization ────────────────────────────────────────────────────────────
matplotlib==3.8.3
seaborn==0.13.2
tabulate==0.9.0

# ── Feature Extraction ───────────────────────────────────────────────────────
joblib==1.3.2

# ── API / Deployment ─────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9
pydantic==2.6.4

# ── ONNX Export ───────────────────────────────────────────────────────────────
onnx==1.15.0
onnxruntime==1.17.1

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm==4.66.2
python-dotenv==1.0.1
PyYAML==6.0.1
scipy==1.12.0
colorama==0.4.6
